Update patch set 1
Patch Set 1:

Merging because the version code depends on the commits we have on top of microG tags, so even if we don't end up using this, we'd have to revert if we want to ship an APK with this

i.e. current APK - older version code
APK with this - two commits extra, higher version code
APK if this is reverted - even higher version code.

Patch-set: 1
Reviewer: Gerrit User 1000000 <1000000@35266860-6252-469c-a67d-185064479fc0>